0728 change Array index to int
Issue: https://gitee.com/openharmony/arkcompiler_runtime_core/issues/ICSH7E

+++ frameworks/bridge/arkts_frontend/koala_projects/arkoala-arkts/arkui-ohos/src/handwritten/Router.ts
@@ -181,8 +181,8 @@ class RouterImpl implements Router {
             return;
         }
         if (this.visiblePages.length > index && this.peerNodeList.length > index) {
-            this.visiblePages.value[index].updatePeerNode(node)
-            RouterExtender.moveCommonUnderPageNode(node.peer.ptr, this.peerNodeList[index])
+            this.visiblePages.value[index as int32].updatePeerNode(node)
+            RouterExtender.moveCommonUnderPageNode(node.peer.ptr, this.peerNodeList[index as int32])
         }
     }
